face-webui-platform/backend/requirements.txt
# FastAPI ML Dashboard Backend
# Python 3.11+

# ── Core web framework ─────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9             # Required for UploadFile / Form

# ── Pydantic v2 ────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1            # BaseSettings with env-file support

# ── WebSockets ─────────────────────────────────────────────────────────────
websockets==12.0
httpx==0.27.0                       # For async HTTP (testing / internal calls)

# ── System performance tracking ────────────────────────────────────────────
psutil==5.9.8

# ── Numerical / image processing ───────────────────────────────────────────
numpy==1.26.4
opencv-python-headless==4.9.0.80    # Headless for server deployment
Pillow==10.3.0                      # Fallback image decode
matplotlib==3.10.1                  # IEEE result chart generation

# ── ML / face detection ────────────────────────────────────────────────────
facenet-pytorch                     # Provides MTCNN face detector

# ── Async / concurrency ────────────────────────────────────────────────────
anyio==4.3.0

# ── Logging / observability ────────────────────────────────────────────────
python-json-logger==2.0.7           # Optional structured JSON logging

# ── Dev / test (not required in production) ───────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
